Implemented BATCH Create with mongoc driver
