+++ CHANGES_NEXT_RELEASE
@@ -5,3 +5,4 @@
 * Issue #280  TRoE for PUT /entities/{entityId}
 * Issue #280  Reimplementation of BATCH Upsert, using the new mongo driver (mongoc) - only used if -experimental is set
 * Issue #280  Reimplementation of BATCH Update, using the new mongo driver (mongoc) - only used if -experimental is set
+* Issue #280  Reimplementation of BATCH Create, using the new mongo driver (mongoc) - only used if -experimental is set
